[#159] :truck: rename: writer info 컴포넌트 디렉토리 변경
