+++ client/src/components/common/diary/WriterInfo.tsx
